# backend-api/.build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GetUserResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string",
						"nullable": true
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string",
						"nullable": true
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetPatientResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"properties": {
							"urlImage": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"cpf",
					"birthDate",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatientRequest": {
				"properties": {
					"idUser": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"idUser",
					"name",
					"cpf",
					"birthDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCaregiverResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"corenNumber": {
						"type": "string"
					},
					"biography": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"properties": {
							"urlImage": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"corenNumber",
					"biography",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CaregiverRequest": {
				"properties": {
					"idUser": {
						"type": "number",
						"format": "double"
					},
					"corenNumber": {
						"type": "string"
					},
					"biography": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"idUser",
					"corenNumber",
					"biography"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AppointmentStatus": {
				"enum": [
					"Scheduled",
					"Completed",
					"Canceled"
				],
				"type": "string"
			},
			"GetAppointmentResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"scheduledAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/AppointmentStatus"
					},
					"patient": {
						"properties": {
							"phoneNumber": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"phoneNumber",
							"name"
						],
						"type": "object"
					},
					"caregiver": {
						"properties": {
							"phoneNumber": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"phoneNumber",
							"name"
						],
						"type": "object"
					},
					"address": {
						"properties": {
							"alias": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"alias",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"scheduledAt",
					"status",
					"patient",
					"caregiver",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostAppointmentRequest": {
				"properties": {
					"scheduledAt": {
						"type": "string",
						"format": "date-time"
					},
					"idPatient": {
						"type": "number",
						"format": "double"
					},
					"idCaregiver": {
						"type": "number",
						"format": "double"
					},
					"idAddress": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"scheduledAt",
					"idPatient",
					"idCaregiver",
					"idAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutAppointmentRequest": {
				"properties": {
					"scheduledAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/AppointmentStatus"
					},
					"idAddress": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"scheduledAt",
					"status",
					"idAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"firebase": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "homecare-backend-api",
		"version": "1.0.0",
		"description": "API para o sistema de agendamento de atendimentos HomeCareApp",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Dados do usuário",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um usuário pelo seu ID",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Altera os dados de um usuário",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Remove o registro de um usuário da base",
				"tags": [
					"Users"
				],
				"security": [
					{
						"firebase": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/patients": {
			"get": {
				"operationId": "GetAllPatients",
				"responses": {
					"200": {
						"description": "Lista de todos os pacientes e seus dados",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetPatientResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Busca a lista de todos os pacientes da base",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePatient",
				"responses": {
					"200": {
						"description": "Retorna o ID do cadastro criado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria um novo cadastro de paciente na base",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatientRequest"
							}
						}
					}
				}
			}
		},
		"/patients/{id}": {
			"get": {
				"operationId": "GetPatientById",
				"responses": {
					"200": {
						"description": "Retorna os dados do paciente consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPatientResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um paciente específico pelo seu ID",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePatient",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Atualiza o cadastro de paciente na base",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PatientRequest"
							}
						}
					}
				}
			}
		},
		"/patients/user/{idUser}": {
			"get": {
				"operationId": "GetPatientByUserId",
				"responses": {
					"200": {
						"description": "Retorna os dados do paciente consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPatientResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um paciente específico pelo ID do usuário associado",
				"tags": [
					"Patients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUser",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/care-professionals": {
			"get": {
				"operationId": "GetAllCareProfessinals",
				"responses": {
					"200": {
						"description": "Lista de todos os cuidadores e seus dados",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetCaregiverResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Busca a lista de todos os cuidadores da base",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCareProfessinal",
				"responses": {
					"200": {
						"description": "Retorna o ID do cadastro criado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria um novo cadastro de cuidador na base",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CaregiverRequest"
							}
						}
					}
				}
			}
		},
		"/care-professionals/{id}": {
			"get": {
				"operationId": "GetCareProfessinalById",
				"responses": {
					"200": {
						"description": "Retorna os dados do cuidador consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCaregiverResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um cuidador específico pelo seu ID",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCareProfessinal",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Atualiza o cadastro de cuidador na base",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CaregiverRequest"
							}
						}
					}
				}
			}
		},
		"/care-professionals/user/{idUser}": {
			"get": {
				"operationId": "GetCareProfessinalByUserId",
				"responses": {
					"200": {
						"description": "Retorna os dados do cuidador consultado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCaregiverResponse"
								}
							}
						}
					}
				},
				"summary": "Busca um cuidador específico pelo ID do usuário associado",
				"tags": [
					"Caregivers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUser",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/appointments": {
			"get": {
				"operationId": "GetAllAppointments",
				"responses": {
					"200": {
						"description": "Lista de todos os agendamentos",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetAppointmentResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Busca por todos os agendamentos da base",
				"tags": [
					"Appointments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "idCaregiver",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "idPatient",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAppointment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"summary": "Cria um novo agendamento",
				"tags": [
					"Appointments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostAppointmentRequest"
							}
						}
					}
				}
			}
		},
		"/appointments/{id}": {
			"get": {
				"operationId": "GetAppointmentById",
				"responses": {
					"200": {
						"description": "Exibe os dados do agendamento",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAppointmentResponse"
								}
							}
						}
					}
				},
				"summary": "Busca por um agendamento pelo seu ID",
				"tags": [
					"Appointments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAppointment",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"summary": "Atualiza um agendamento pelo ID",
				"tags": [
					"Appointments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutAppointmentRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAppointment",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"summary": "Remove um agendamento da base",
				"tags": [
					"Appointments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}